refactor: Rename `plot_tools.py` to `manager.py`
