+++ chia/plotting/manager.py
